#10 Finished
- Remove the need of "`Export -> Resources -> Filters to export non-resources...`, you need to put this: `res://addons/admob/*.json`"
- Changed the `.json` to a `.gd` Script which extends Node and only have a Dictionary Variable to store the config variable, located at: `addons/admob/config/Settings.gd`
- Removed `addons/admob/config/settings.json` due is useless now with the `Settings,gd`

Note: If change the Variables directly from the Editor and change the screen, then will give this error: ` Another resource is loaded from path 'res://addons/admob/config/Settings.gd' (possible cyclic resource inclusion).
`  because the Script is not fully reloaded when is saved, this won't change the functionality of the Plugin so it will only need to fix this for next versions.